WIP refactor java FFI code
